trace: hook plugin to Flexus API
This finalize the plugin coupling with the Flexus API.
The 'decoders' were imported nearly as is from past version
of libqflex.

The commit and implementation may contain some flaws.
Regarding the physical_address_t, it's rather hard to
undetstand which physical address is expected for
the instruction or memory dispatch, and whether the offcial
API return what is actually needed.